refactor: lottery 일반 type - api type 분리
