+++ admin/src/hooks/useInfiniteFetch.ts
@@ -1,5 +1,4 @@
 import { useCallback, useEffect, useState } from "react";
-import { flushSync } from "react-dom";
 import { InfiniteListData } from "@/types/common";
 
 interface UseInfiniteFetchProps<R> {
