+++ admin/src/types/lottery.ts
@@ -1,38 +1,15 @@
-import { InfiniteListData } from "./common";
-
-export type GetLotteryResponse = {
+export interface LotteryType {
     lotteryEventId: number;
     startDate: string;
     endDate: string;
     appliedCount: number;
     winnerCount: number;
-}[];
-
-export interface PostLotteryWinnerParams {
-    id: number;
-}
-
-export interface PostLotteryWinnerResponse {
-    message: string;
-}
-
-export interface GetLotteryWinnerParams {
-    id: number;
-    size: number;
-    page: number;
-}
-
-export interface GetLotteryExpectationsParams {
-    lotteryId: number;
-    participantId: number;
 }
 
 export interface LotteryExpectationsType {
     casperId: number;
     expectation: string;
 }
-
-export type GetLotteryExpectationsResponse = LotteryExpectationsType[];
 
 export interface LotteryWinnerType {
     id: number;
@@ -41,4 +18,3 @@ export interface LotteryWinnerType {
     expectation: number;
     appliedCount: number;
 }
-export interface GetLotteryWinnerResponse extends InfiniteListData<LotteryWinnerType> {}
